+++ services/src/main/java/by/pvt/kish/aircompany/services/impl/FlightService.java
@@ -1,14 +1,14 @@
 package by.pvt.kish.aircompany.services.impl;
 
+import by.pvt.kish.aircompany.constants.Message;
 import by.pvt.kish.aircompany.dao.impl.FlightDAO;
 import by.pvt.kish.aircompany.entity.Flight;
 import by.pvt.kish.aircompany.exceptions.DaoException;
-import by.pvt.kish.aircompany.exceptions.ServiceValidateException;
 import by.pvt.kish.aircompany.exceptions.ServiceException;
+import by.pvt.kish.aircompany.exceptions.ServiceValidateException;
 import by.pvt.kish.aircompany.services.BaseService;
 import by.pvt.kish.aircompany.validators.FlightValidator;
 
-import java.sql.SQLException;
 import java.util.List;
 
 /**
@@ -35,29 +35,50 @@ public class FlightService extends BaseService<Flight> {
             }
             return flightDAO.add(flight);
         } catch (DaoException e) {
-            throw new ServiceException("Creation flight failed", e);
+            throw new ServiceException(e.getMessage());
         }
     }
 
     @Override
-    public void update(Flight flight) throws SQLException {
-        flightDAO.update(flight);
+    public void update(Flight flight) throws ServiceException, ServiceValidateException {
+        try {
+            String validateResult = FlightValidator.validate(flight);
+            if (validateResult != null) {
+                throw new ServiceValidateException(validateResult);
+            }
+            flightDAO.update(flight);
+        } catch (DaoException e) {
+            throw new ServiceException(e.getMessage());
+        }
     }
 
     @Override
-    public List<Flight> getAll() throws SQLException {
-        List<Flight> list = flightDAO.getAll();
-        return list;
+    public List<Flight> getAll() throws ServiceException {
+        try {
+            return flightDAO.getAll();
+        }catch (DaoException e) {
+            throw new ServiceException(e.getMessage());
+        }
     }
 
     @Override
-    public void delete(int id) throws SQLException {
-        flightDAO.delete(id);
+    public void delete(int id) throws ServiceException {
+        try {
+            if (id < 0) {
+                throw new ServiceException(Message.ERROR_ID_MISSING);
+            }
+            flightDAO.delete(id);
+        } catch (DaoException e) {
+            throw new ServiceException(e.getMessage());
+        }
     }
 
     @Override
-    public Flight getById(int id) throws SQLException {
-        Flight flight = flightDAO.getById(id);
-        return flight;
+    public Flight getById(int id) throws ServiceException {
+        try {
+            return flightDAO.getById(id);
+        } catch (DaoException e) {
+            throw new ServiceException(e.getMessage());
+        }
     }
 }
